adapt code to new backend location
